requirements.txt
customtkinter==5.2.2
matplotlib==3.10.8
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
polars==1.36.1
polars-runtime-32==1.36.1
PyYAML==6.0.3
rebrick==0.4.0
requests==2.32.5
scikit-image==0.26.0
scipy==1.16.3
tkinter-async-execute==1.4.0
torch==2.9.1
torchvision==0.24.1
ultralytics==8.3.250
